subscriber1/requirements.txt
azure-functions
azure-servicebus
azure-storage-blob